Add support for fb initializers in functions, moved tests around
