+++ src/codegen/tests/initialization_test.rs
@@ -1,0 +1,3 @@
+mod global_initializers;
+mod pou_initializers;
+mod type_initializers;
